Modificamos el formato del proyecto para usar Maven
