--- src/Modelo/Errores/CasilleroOcupado.java
@@ -1,4 +1,0 @@
-package Modelo.Errores;
-
-public class CasilleroOcupado extends RuntimeException {
-}
