--- src/Modelo/Juego/Main.java
@@ -1,10 +1,0 @@
-package Modelo.Juego;
-
-
-public class Main {
-
-	public static void main(){
-
-		
-	}
-}
