--- src/Modelo/Unidad/Entidad.java
@@ -1,12 +1,0 @@
-package Modelo.Unidad;
-
-
-import Modelo.Tablero.Casillero;
-
-public abstract class Entidad {
-    protected Casillero casillero;
-
-    public void setCasillero(Casillero casillero){
-        this.casillero = casillero;
-    }
-}
