+++ src/main/java/fiuba/AlgoChess/Modelo/Errores/CasilleroOcupadoException.java
@@ -1,0 +1,4 @@
+package fiuba.AlgoChess.Modelo.Errores;
+
+public class CasilleroOcupadoException extends RuntimeException {
+}
