+++ src/main/java/fiuba/AlgoChess/Modelo/Juego/Main.java
@@ -1,0 +1,8 @@
+package fiuba.AlgoChess.Modelo.Juego;
+
+public class Main {
+
+    public static void main(String[] args) {
+
+    }
+}
