+++ src/main/java/fiuba/AlgoChess/Modelo/Tablero/Estado.java
@@ -1,0 +1,8 @@
+package fiuba.AlgoChess.Modelo.Tablero;
+
+import fiuba.AlgoChess.Modelo.Unidad.Entidad;
+
+public abstract class Estado {
+
+    public abstract void agregarEntidad(Entidad entidad, Casillero casillero);
+}
