+++ src/main/java/fiuba/AlgoChess/Modelo/Tablero/Libre.java
@@ -1,10 +1,12 @@
-package Modelo.Tablero;
+package fiuba.AlgoChess.Modelo.Tablero;
 
-import Modelo.Unidad.Entidad;
+import fiuba.AlgoChess.Modelo.Unidad.Entidad;
 
 public class Libre extends Estado {
+
     @Override
     public void agregarEntidad(Entidad entidad, Casillero casillero) {
+
         entidad.setCasillero(casillero);
     }
 }
