+++ src/main/java/fiuba/AlgoChess/Modelo/Tablero/Ocupado.java
@@ -1,0 +1,14 @@
+package fiuba.AlgoChess.Modelo.Tablero;
+
+import fiuba.AlgoChess.Modelo.Errores.CasilleroOcupadoException;
+import fiuba.AlgoChess.Modelo.Unidad.Entidad;
+
+public class Ocupado extends Estado {
+
+    @Override
+    public void agregarEntidad(Entidad entidad, Casillero casillero){
+
+        throw new CasilleroOcupadoException();
+    }
+}
+
