+++ src/main/java/fiuba/AlgoChess/Modelo/Unidad/Entidad.java
@@ -1,0 +1,18 @@
+package fiuba.AlgoChess.Modelo.Unidad;
+
+import fiuba.AlgoChess.Modelo.Tablero.Casillero;
+
+public abstract class Entidad {
+
+    private Casillero casillero;
+
+    public Entidad(Casillero casillero) {
+
+        this.setCasillero(casillero);
+    }
+
+    public void setCasillero(Casillero casillero){
+
+        this.casillero = casillero;
+    }
+}
