+++ src/main/java/fiuba/AlgoChess/Modelo/Unidad/Jinete.java
@@ -1,16 +1,16 @@
-package Modelo.Unidad;
+package fiuba.AlgoChess.Modelo.Unidad;
 
-import Modelo.Tablero.Casillero;
+import fiuba.AlgoChess.Modelo.Tablero.Casillero;
 
 public class Jinete extends Entidad {
+
     private int vida = 100;
     private int danioDistancia = 15;
     private int danioMelee = 5;
 
     public Jinete(Casillero casillero){
-        this.casillero = casillero;
+
+        super(casillero);
     }
-
-
 
 }
