+++ src/test/java/fiuba/AlgoChess/Modelo/Tablero/CasilleroTest.java
@@ -1,0 +1,15 @@
+package fiuba.AlgoChess.Modelo.Tablero;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class CasilleroTest {
+
+    // Test de ejemplo con JUnit 4, borrar luego.
+    @Test
+    public void unTest() {
+
+        Assert.assertEquals(2, 2);
+    }
+
+}
