--- test/Modelo/Tablero/CasilleroTest.java
@@ -1,16 +1,0 @@
-package Modelo.Tablero;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Test; // Estoy usando JUnit 5
-
-
-public class CasilleroTest {
-	
-	// Test de ejemplo con JUnit 5, borrar luego.
-	@Test
-	void unTest() {
-		
-		assertEquals(2, 2);
-	}
-
-}
